project/requirements.txt
flask==2.0.1
flask-cors==3.0.10
face-recognition==1.3.0
numpy==1.21.0
Pillow==8.3.1
pymysql==1.0.2 